requirements.txt
fastapi
pydantic
PyJWT
uvicorn
